requirements.txt
python-dotenv
pandas
numpy
pymongo
scikit-learn

-e .
